refactor(runners): extract HealthCheckFallback concern and fix codex stdin for string commands
Address two review threads on PR #2692:

1. Extract six near-identical health-check fallback methods from
   Runners::TestAgent and Providers::TestAgent into a shared
   TestAgentHealthCheckFallback concern. Each class implements three
   abstract methods (harness_health_check_key, harness_fallback_log_prefix,
   harness_fallback_log_context) so future changes to fallback patterns or
   the detection regex only need to happen in one place.

2. Move close_stdin_for_codex_exec before the sh -c wrapping in
   execute_unlocked so it operates on the original command. Update the
   method to handle both string and array-form commands via
   Shellwords.split, consistent with how codex_auth_lock_required? already
   operates on the pre-wrapping command.
