+++ app/services/runners/test_agent.rb
@@ -255,51 +255,22 @@ module Runners
       Result.new(success: true, error_type: nil, message: "Diagnostic passed (#{summary})")
     end
 
+    include TestAgentHealthCheckFallback
+
+    def harness_health_check_key
+      harness_runner_key
+    end
+
+    def harness_fallback_log_prefix
+      "runners.test_agent"
+    end
+
+    def harness_fallback_log_context
+      { runner_key: runner.runner_key }
+    end
+
     def execute_harness_health_check
       AgentHarness.check_provider(harness_runner_key, timeout: TIMEOUT)
-    end
-
-    def execute_harness_health_check_with_fallback
-      process_harness_result(maybe_fallback_harness_result(execute_harness_health_check))
-    rescue StandardError => e
-      raise unless fallback_to_container_smoke_test?(e.message)
-
-      log_harness_fallback(error_message: e.message, error_class: e.class.name)
-      execute_container_smoke_test
-    end
-
-    def maybe_fallback_harness_result(result)
-      return result unless fallback_to_container_smoke_test?(result[:message], result[:output])
-
-      log_harness_fallback(error_message: [ result[:message], result[:output] ].compact.join("\n"))
-      execute_container_smoke_test_raw
-    end
-
-    def execute_container_smoke_test_raw
-      test_run = build_test_run
-
-      begin
-        test_run.with_container do |run|
-          executor = Containers::HarnessExecutor.new(run)
-          prepare_kilocode_config!(run) if kilocode_direct_outbound?
-          AgentHarness.check_provider(
-            harness_runner_key,
-            timeout: TIMEOUT,
-            executor: executor,
-            provider_runtime: container_provider_runtime
-          )
-        end
-      ensure
-        test_run.destroy! if test_run&.persisted?
-      end
-    end
-
-    # Runs the agent-harness smoke_test contract inside a provisioned container.
-    #
-    # Instead of building provider-specific CLI commands locally, this delegates
-    # to the harness provider's smoke_test method with a container-backed executor.
-    def execute_container_smoke_test
-      process_harness_result(execute_container_smoke_test_raw)
     end
 
     def process_harness_result(result)
@@ -530,26 +501,6 @@ module Runners
       return false if runner.subscription?
       api_key_name = RunnerSupport.proxy_health_check_api_key_for(runner.runner_key)
       !!(api_key_name && proxy_api_key_configured?(api_key_name))
-    end
-
-    def fallback_to_container_smoke_test?(*messages)
-      return false unless test_project
-
-      combined = messages.compact.map { |message| normalize_output_text(message) }.join("\n")
-      return false if combined.blank?
-
-      combined.match?(/permission denied \(os error 13\)/i) ||
-        combined.match?(/sandbox failure detected/i) ||
-        combined.match?(/bwrap:.*permission denied/i)
-    end
-
-    def log_harness_fallback(error_message:, error_class: nil)
-      Rails.logger.warn(
-        message: "runners.test_agent.host_health_check_fallback",
-        runner_key: runner.runner_key,
-        error_class: error_class,
-        error_message: normalize_output_text(error_message)
-      )
     end
 
     def harness_runner_key
